+++ src/daemon/handlers/__tests__/interaction-touch.test.ts
@@ -1,0 +1,188 @@
+import { test, expect, vi, beforeEach } from 'vitest';
+import { attachRefs } from '@agent-device/kernel/snapshot';
+import { makeSessionStore } from '../../../__tests__/test-utils/store-factory.ts';
+import { handleInteractionCommands } from '../interaction.ts';
+import {
+  contextFromFlags,
+  createEmulateCaptureSnapshotForSession,
+  makeMacOsDesktopSession,
+  makeMacOsMenubarSession,
+  makeSession,
+} from './interaction-touch-fixtures.ts';
+
+// Router ownership: one representative per touch command proves
+// handleTouchInteractionCommands claims press/click/longpress/fill.
+
+vi.mock('../../../core/dispatch.ts', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('../../../core/dispatch.ts')>();
+  return {
+    ...actual,
+    dispatchCommand: vi.fn(async () => ({})),
+  };
+});
+
+vi.mock('../interaction-snapshot.ts', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('../interaction-snapshot.ts')>();
+  return {
+    ...actual,
+    captureSnapshotForSession: vi.fn(async () => ({
+      nodes: [],
+      createdAt: 0,
+      backend: 'xctest' as const,
+    })),
+  };
+});
+
+import { dispatchCommand } from '../../../core/dispatch.ts';
+const mockDispatch = vi.mocked(dispatchCommand);
+import { captureSnapshotForSession } from '../interaction-snapshot.ts';
+const mockCaptureSnapshotForSession = vi.mocked(captureSnapshotForSession);
+
+beforeEach(() => {
+  mockDispatch.mockReset();
+  mockDispatch.mockResolvedValue({});
+  mockCaptureSnapshotForSession.mockReset();
+  mockCaptureSnapshotForSession.mockImplementation(
+    createEmulateCaptureSnapshotForSession(mockDispatch),
+  );
+});
+
+test('press coordinates dispatches press and records as press', async () => {
+  const sessionStore = makeSessionStore();
+  const sessionName = 'default';
+  const storedSession = makeSession(sessionName);
+  sessionStore.set(sessionName, storedSession);
+
+  mockDispatch.mockResolvedValue({ ok: true });
+
+  const response = await handleInteractionCommands({
+    req: {
+      token: 't',
+      session: sessionName,
+      command: 'press',
+      positionals: ['100', '200'],
+      flags: { count: 3, intervalMs: 1, doubleTap: true },
+    },
+    sessionName,
+    sessionStore,
+    contextFromFlags,
+  });
+
+  expect(response).toBeTruthy();
+  expect(response?.ok).toBe(true);
+  expect(mockDispatch).toHaveBeenCalledTimes(1);
+  expect(mockDispatch.mock.calls[0]?.[1]).toBe('press');
+  expect(mockDispatch.mock.calls[0]?.[2]).toEqual(['100', '200']);
+  const context = mockDispatch.mock.calls[0]?.[4] as Record<string, unknown> | undefined;
+  expect(context?.count).toBe(3);
+  expect(context?.intervalMs).toBe(1);
+  expect(context?.doubleTap).toBe(true);
+
+  const session = sessionStore.get(sessionName);
+  expect(session).toBeTruthy();
+  expect(session?.actions.length).toBe(1);
+  expect(session?.actions[0]?.command).toBe('press');
+  expect(session?.actions[0]?.positionals).toEqual(['100', '200']);
+});
+
+test('click rejects macOS desktop surface interactions until helper routing exists', async () => {
+  const sessionStore = makeSessionStore();
+  const sessionName = 'macos-desktop-click';
+  sessionStore.set(sessionName, makeMacOsDesktopSession(sessionName));
+
+  mockDispatch.mockRejectedValue(new Error('dispatch should not be called'));
+
+  const response = await handleInteractionCommands({
+    req: {
+      token: 't',
+      session: sessionName,
+      command: 'click',
+      positionals: ['100', '200'],
+      flags: {},
+    },
+    sessionName,
+    sessionStore,
+    contextFromFlags,
+  });
+
+  expect(response?.ok).toBe(false);
+  if (response && !response.ok) {
+    expect(response.error.code).toBe('UNSUPPORTED_OPERATION');
+    expect(response.error.message).toMatch(/macOS desktop sessions/);
+  }
+});
+
+test('fill rejects macOS menubar surface interactions until helper routing exists', async () => {
+  const sessionStore = makeSessionStore();
+  const sessionName = 'macos-menubar-fill';
+  sessionStore.set(sessionName, makeMacOsMenubarSession(sessionName));
+
+  mockDispatch.mockRejectedValue(new Error('dispatch should not be called'));
+
+  const response = await handleInteractionCommands({
+    req: {
+      token: 't',
+      session: sessionName,
+      command: 'fill',
+      positionals: ['@e2', 'hello'],
+      flags: {},
+    },
+    sessionName,
+    sessionStore,
+    contextFromFlags,
+  });
+
+  expect(response?.ok).toBe(false);
+  if (response && !response.ok) {
+    expect(response.error.code).toBe('UNSUPPORTED_OPERATION');
+    expect(response.error.message).toMatch(/macOS menubar sessions/);
+  }
+});
+
+test('longpress @ref resolves the target and dispatches coordinate longpress', async () => {
+  const sessionStore = makeSessionStore();
+  const sessionName = 'longpress-ref';
+  const session = makeSession(sessionName);
+  session.snapshot = {
+    nodes: attachRefs([
+      {
+        index: 0,
+        type: 'XCUIElementTypeStaticText',
+        label: 'Last message',
+        rect: { x: 10, y: 20, width: 100, height: 40 },
+        enabled: true,
+        hittable: true,
+      },
+    ]),
+    createdAt: Date.now(),
+    backend: 'xctest',
+  };
+  sessionStore.set(sessionName, session);
+
+  mockDispatch.mockResolvedValue({ native: true });
+
+  const response = await handleInteractionCommands({
+    req: {
+      token: 't',
+      session: sessionName,
+      command: 'longpress',
+      positionals: ['@e1', '800'],
+      flags: {},
+    },
+    sessionName,
+    sessionStore,
+    contextFromFlags,
+  });
+
+  expect(response?.ok).toBe(true);
+  if (response?.ok) {
+    expect(response.data?.x).toBe(60);
+    expect(response.data?.y).toBe(40);
+    expect(response.data?.durationMs).toBe(800);
+    expect(response.data?.message).toMatch(/Long pressed @e1/);
+  }
+  expect(mockDispatch).toHaveBeenCalledTimes(1);
+  expect(mockDispatch.mock.calls[0]?.[1]).toBe('longpress');
+  expect(mockDispatch.mock.calls[0]?.[2]).toEqual(['60', '40', '800']);
+  expect(sessionStore.get(sessionName)?.actions[0]?.command).toBe('longpress');
+});
