+++ src/daemon/handlers/interaction-touch-press.ts
@@ -22,7 +22,10 @@ import {
 import { refMutationAdmissionResponse } from './interaction-ref-policy.ts';
 import { createInteractionRuntime } from './interaction-runtime.ts';
 import type { CaptureSnapshotForSession } from './interaction-snapshot.ts';
-import { refreshAndroidRefSnapshotIfFreshnessActive } from './interaction-touch-android-readiness.ts';
+import {
+  refreshAndroidRefSnapshotIfFreshnessActive,
+  type RefAdmissionContext,
+} from './interaction-touch-android-readiness.ts';
 import {
   dispatchDirectIosSelectorTap,
   readDirectIosSelectorTapTarget,
@@ -38,6 +41,8 @@ import {
   formatTouchTargetLabel,
   parseLongPressTarget,
   parseTouchTarget,
+  type ParsedLongPressTarget,
+  type ParsedTouchTarget,
 } from './interaction-touch-targets.ts';
 import { errorResponse, noActiveSessionError, requireCommandSupported } from './response.ts';
 
@@ -49,18 +54,114 @@ import { errorResponse, noActiveSessionError, requireCommandSupported } from './
 
 export type TargetedTouchCommand = 'press' | 'click' | 'longpress';
 
+/** The label the user typed, which capability policy and parse errors quote. */
+type TargetedTouchCommandLabel = TargetedTouchCommand;
+
+type TargetedTouchParams = InteractionHandlerParams & {
+  captureSnapshotForSession: CaptureSnapshotForSession;
+  refSnapshotFlagGuardResponse: RefSnapshotFlagGuardResponse;
+};
+
+/** What survives admission: the target this dispatch acts on and its policy context. */
+type AdmittedTargetedTouch = {
+  session: SessionState;
+  commandLabel: TargetedTouchCommandLabel;
+  clickButton: ReturnType<typeof resolveClickButton>;
+  resultButtonTag: ReturnType<typeof buttonTag>;
+  target: InteractionTarget;
+  durationMs: number | undefined;
+  staleRefsWarning: string | undefined;
+  androidFreshnessBaseline: SessionState['snapshot'];
+  refContext: RefAdmissionContext | undefined;
+};
+
 export async function dispatchTargetedTouchViaRuntime(
-  params: InteractionHandlerParams & {
-    captureSnapshotForSession: CaptureSnapshotForSession;
-    refSnapshotFlagGuardResponse: RefSnapshotFlagGuardResponse;
-  },
+  params: TargetedTouchParams,
   command: TargetedTouchCommand,
 ): Promise<DaemonResponse> {
-  const { req, sessionName, sessionStore } = params;
+  const admission = await admitTargetedTouch(params, command);
+  if ('response' in admission) return admission.response;
+  const { admitted } = admission;
+
+  // ADR 0012 step 4: a guarded replay dispatch must resolve through the
+  // runtime tree path so the post-resolution identity guard runs — the
+  // direct-iOS fast path has no daemon-tree node to check against.
+  const directSelector = params.req.internal?.replayTargetGuard
+    ? null
+    : readDirectIosSelectorTapTarget({
+        session: admitted.session,
+        commandLabel: admitted.commandLabel,
+        target: admitted.target,
+        flags: params.req.flags,
+      });
+  if (directSelector) {
+    const directResponse = await dispatchDirectIosSelectorTap(
+      params,
+      admitted.session,
+      directSelector,
+    );
+    if (directResponse) return directResponse;
+  }
+
+  return await dispatchRuntimeInteraction(
+    params,
+    buildTargetedRuntimeOptions(params, command, admitted),
+  );
+}
+
+type ParsedTargetedTouch = Extract<ParsedTouchTarget | ParsedLongPressTarget, { ok: true }>;
+
+async function admitTargetedTouch(
+  params: TargetedTouchParams,
+  command: TargetedTouchCommand,
+): Promise<{ response: DaemonResponse } | { admitted: AdmittedTargetedTouch }> {
+  const { req, sessionStore, sessionName } = params;
   const session = sessionStore.get(sessionName);
-  if (!session) return noActiveSessionError();
+  if (!session) return { response: noActiveSessionError() };
 
   const commandLabel = command === 'click' ? 'click' : command;
+  const policyResponse = targetedTouchPolicyResponse(session, command, commandLabel, req.flags);
+  if (policyResponse) return { response: policyResponse };
+
+  const positionals = req.positionals ?? [];
+  const parsedTarget =
+    command === 'longpress'
+      ? parseLongPressTarget(positionals)
+      : parseTouchTarget(positionals, commandLabel);
+  if (!parsedTarget.ok) return { response: parsedTarget.response };
+  const staleRefsWarning = readTargetedTouchStalenessWarning(session, req, parsedTarget);
+  const refAdmission = await admitTargetedTouchRef(
+    params,
+    session,
+    command,
+    parsedTarget,
+    staleRefsWarning,
+  );
+  if (refAdmission.response) return { response: refAdmission.response };
+
+  const clickButton = resolveClickButton(req.flags);
+  return {
+    admitted: {
+      session,
+      commandLabel,
+      clickButton,
+      resultButtonTag: buttonTag(clickButton),
+      target: parsedTarget.target,
+      durationMs: command === 'longpress' ? parsedTarget.durationMs : undefined,
+      staleRefsWarning,
+      androidFreshnessBaseline: refAdmission.androidFreshnessBaseline,
+      refContext: readRefAdmissionContext(req, parsedTarget, staleRefsWarning),
+    },
+  };
+}
+
+/** Surface, capability, settle, and click-button policy, in that order. */
+function targetedTouchPolicyResponse(
+  session: SessionState,
+  command: TargetedTouchCommand,
+  commandLabel: TargetedTouchCommandLabel,
+  flags: CommandFlags | undefined,
+): DaemonResponse | undefined {
   const capabilityCommand = command === 'longpress' ? 'longpress' : 'press';
   const unsupportedSurfaceResponse = unsupportedMacOsDesktopSurfaceInteraction(
     session,
@@ -69,125 +170,132 @@ export async function dispatchTargetedTouchViaRuntime(
   if (unsupportedSurfaceResponse) return unsupportedSurfaceResponse;
   const unsupported = requireCommandSupported(capabilityCommand, session.device);
   if (unsupported) return unsupported;
-  const invalidSettleFlags = settleFlagGuardResponse(command, req.flags);
+  const invalidSettleFlags = settleFlagGuardResponse(command, flags);
   if (invalidSettleFlags) return invalidSettleFlags;
+  return clickButtonValidationResponse(session, command, commandLabel, flags);
+}
 
-  const clickButton = resolveClickButton(req.flags);
-  const resultButtonTag = buttonTag(clickButton);
-  if (command !== 'longpress' && clickButton !== 'primary') {
-    const validationError = getClickButtonValidationError({
-      commandLabel,
-      platform: publicPlatformString(session.device),
-      button: clickButton,
-      count: req.flags?.count,
-      intervalMs: req.flags?.intervalMs,
-      holdMs: req.flags?.holdMs,
-      jitterPx: req.flags?.jitterPx,
-      doubleTap: req.flags?.doubleTap,
-    });
-    if (validationError) {
-      return errorResponse(validationError.code, validationError.message, validationError.details);
-    }
-  }
+function clickButtonValidationResponse(
+  session: SessionState,
+  command: TargetedTouchCommand,
+  commandLabel: TargetedTouchCommandLabel,
+  flags: CommandFlags | undefined,
+): DaemonResponse | undefined {
+  const clickButton = resolveClickButton(flags);
+  if (command === 'longpress' || clickButton === 'primary') return undefined;
+  const validationError = getClickButtonValidationError({
+    commandLabel,
+    platform: publicPlatformString(session.device),
+    button: clickButton,
+    count: flags?.count,
+    intervalMs: flags?.intervalMs,
+    holdMs: flags?.holdMs,
+    jitterPx: flags?.jitterPx,
+    doubleTap: flags?.doubleTap,
+  });
+  if (!validationError) return undefined;
+  return errorResponse(validationError.code, validationError.message, validationError.details);
+}
 
-  const parsedTarget =
-    command === 'longpress'
-      ? parseLongPressTarget(req.positionals ?? [])
-      : parseTouchTarget(req.positionals ?? [], commandLabel);
-  if (!parsedTarget.ok) return parsedTarget.response;
-  // Staleness relative to what the client knew when it sent this @ref — read
-  // BEFORE any internal recapture (Android freshness refresh, --verify) advances
-  // the generation as a side effect of this same command. Pinned refs
-  // (`@e12~s3`) get a precise generation-mismatch warning; a plain ref warns
-  // while the frame is expired. A mutating `find`'s internal dispatch supplies a
-  // locator-minted ref (`internal.findResolvedTarget`), so it carries no
-  // user-facing staleness — the caller never consumed a `@ref` (ADR 0014).
-  const staleRefsWarning =
-    parsedTarget.target.kind === 'ref' && req.internal?.findResolvedTarget === undefined
-      ? resolveRefStalenessWarning({
-          session,
-          ref: parsedTarget.target.ref,
-          mintedGeneration: parsedTarget.refGeneration,
-        })
-      : undefined;
-  let androidFreshnessBaseline: SessionState['snapshot'];
-  if (parsedTarget.target.kind === 'ref') {
-    const invalidRefFlagsResponse = params.refSnapshotFlagGuardResponse(
-      command === 'longpress' ? 'longpress' : 'press',
-      req.flags,
-    );
-    if (invalidRefFlagsResponse) return invalidRefFlagsResponse;
-    const admissionResponse = req.internal?.findResolvedTarget
-      ? null
-      : refMutationAdmissionResponse({
-          session,
-          ref: parsedTarget.target.ref,
-          mintedGeneration: parsedTarget.refGeneration,
-          staleRefsWarning,
-        });
-    if (admissionResponse) return admissionResponse;
-    androidFreshnessBaseline = await refreshAndroidRefSnapshotIfFreshnessActive(params, session);
-  }
-  // ADR 0012 step 4: a guarded replay dispatch must resolve through the
-  // runtime tree path so the post-resolution identity guard runs — the
-  // direct-iOS fast path has no daemon-tree node to check against.
-  const replayTargetGuard = req.internal?.replayTargetGuard;
-  const directSelector = replayTargetGuard
+/**
+ * Staleness relative to what the client knew when it sent this @ref — read
+ * BEFORE any internal recapture (Android freshness refresh, --verify) advances
+ * the generation as a side effect of this same command. Pinned refs
+ * (`@e12~s3`) get a precise generation-mismatch warning; a plain ref warns
+ * while the frame is expired. A mutating `find`'s internal dispatch supplies a
+ * locator-minted ref (`internal.findResolvedTarget`), so it carries no
+ * user-facing staleness — the caller never consumed a `@ref` (ADR 0014).
+ */
+function readTargetedTouchStalenessWarning(
+  session: SessionState,
+  req: InteractionHandlerParams['req'],
+  parsedTarget: ParsedTargetedTouch,
+): string | undefined {
+  if (parsedTarget.target.kind !== 'ref') return undefined;
+  if (req.internal?.findResolvedTarget !== undefined) return undefined;
+  return resolveRefStalenessWarning({
+    session,
+    ref: parsedTarget.target.ref,
+    mintedGeneration: parsedTarget.refGeneration,
+  });
+}
+
+function readRefAdmissionContext(
+  req: InteractionHandlerParams['req'],
+  parsedTarget: ParsedTargetedTouch,
+  staleRefsWarning: string | undefined,
+): RefAdmissionContext | undefined {
+  if (parsedTarget.target.kind !== 'ref') return undefined;
+  if (req.internal?.findResolvedTarget !== undefined) return undefined;
+  return {
+    ref: parsedTarget.target.ref,
+    mintedGeneration: parsedTarget.refGeneration,
+    staleRefsWarning,
+  };
+}
+
+async function admitTargetedTouchRef(
+  params: TargetedTouchParams,
+  session: SessionState,
+  command: TargetedTouchCommand,
+  parsedTarget: ParsedTargetedTouch,
+  staleRefsWarning: string | undefined,
+): Promise<{ response?: DaemonResponse; androidFreshnessBaseline?: SessionState['snapshot'] }> {
+  const { req } = params;
+  if (parsedTarget.target.kind !== 'ref') return {};
+  const invalidRefFlagsResponse = params.refSnapshotFlagGuardResponse(
+    command === 'longpress' ? 'longpress' : 'press',
+    req.flags,
+  );
+  if (invalidRefFlagsResponse) return { response: invalidRefFlagsResponse };
+  const admissionResponse = req.internal?.findResolvedTarget
     ? null
-    : readDirectIosSelectorTapTarget({
+    : refMutationAdmissionResponse({
         session,
-        commandLabel,
-        target: parsedTarget.target,
-        flags: req.flags,
+        ref: parsedTarget.target.ref,
+        mintedGeneration: parsedTarget.refGeneration,
+        staleRefsWarning,
       });
-  if (directSelector) {
-    const directResponse = await dispatchDirectIosSelectorTap(params, session, directSelector);
-    if (directResponse) return directResponse;
-  }
-  const durationMs = command === 'longpress' ? parsedTarget.durationMs : undefined;
+  if (admissionResponse) return { response: admissionResponse };
+  return {
+    androidFreshnessBaseline: await refreshAndroidRefSnapshotIfFreshnessActive(params, session),
+  };
+}
 
-  return await dispatchRuntimeInteraction(params, {
-    androidFreshnessBaseline,
-    refContext:
-      parsedTarget.target.kind === 'ref' && req.internal?.findResolvedTarget === undefined
-        ? {
-            ref: parsedTarget.target.ref,
-            mintedGeneration: parsedTarget.refGeneration,
-            staleRefsWarning,
-          }
-        : undefined,
-    iosTapCorroboration: {
-      target: parsedTarget.target,
-      extra:
-        command === 'longpress'
-          ? {
-              ...(durationMs !== undefined ? { durationMs } : {}),
-              gesture: 'longpress',
-            }
-          : resultButtonTag,
-    },
+function buildTargetedRuntimeOptions(
+  params: TargetedTouchParams,
+  command: TargetedTouchCommand,
+  admitted: AdmittedTargetedTouch,
+): Parameters<typeof dispatchRuntimeInteraction<TargetedTouchResult>>[1] {
+  const { req, sessionName } = params;
+  const { session, target, durationMs, staleRefsWarning, resultButtonTag } = admitted;
+  const targetedExtra =
+    command === 'longpress'
+      ? { ...(durationMs !== undefined ? { durationMs } : {}), gesture: 'longpress' }
+      : resultButtonTag;
+  return {
+    androidFreshnessBaseline: admitted.androidFreshnessBaseline,
+    refContext: admitted.refContext,
+    iosTapCorroboration: { target, extra: targetedExtra },
     run: async (runtime) =>
       await runTargetedTouchInteraction({
         runtime,
         command,
-        target: parsedTarget.target,
+        target,
         sessionName,
         requestId: req.meta?.requestId,
-        clickButton,
+        clickButton: admitted.clickButton,
         flags: req.flags,
         durationMs,
-        expectedResolvedTarget: replayTargetGuard,
+        expectedResolvedTarget: req.internal?.replayTargetGuard,
         preresolvedTarget: req.internal?.findResolvedTarget,
       }),
     afterRun: async (result) => {
       if (session.lease?.leaseProvider) return undefined;
-      return await assertAndroidPressStayedInApp(
-        session,
-        formatTouchTargetLabel(parsedTarget.target, result),
-      );
+      return await assertAndroidPressStayedInApp(session, formatTouchTargetLabel(target, result));
     },
     buildPayloads: async (result) => {
-      const durationMs = readLongPressResultDuration(result);
+      const resultDurationMs = readLongPressResultDuration(result);
       return await buildTargetedTouchResponsePayloads({
         params,
         session,
@@ -202,13 +310,13 @@ export async function dispatchTargetedTouchViaRuntime(
         extra:
           command === 'longpress'
             ? {
-                ...(durationMs !== undefined ? { durationMs } : {}),
+                ...(resultDurationMs !== undefined ? { durationMs: resultDurationMs } : {}),
                 gesture: 'longpress',
               }
             : resultButtonTag,
       });
     },
-  });
+  };
 }
 
 async function runTargetedTouchInteraction(params: {
